refactor(check): extract records.ts from checker.ts
Extract record literal and nested record processing functions into a
dedicated records.ts module (Layer 3 - Constructs). This includes:

- Field init detection: getFieldInitFromLine, getIndentLevelFromLine,
  hasNestedRecordInit, getNestedRecordTypeName, getNestedRecordInitFromFieldInit
- Field init processing: processFieldInitAsTypeField, processFieldInitInNestedContext
- Record literal: startRecordLiteral, extractFieldInitName, finalizeRecordLiteral
- Nested record init: startNestedRecordInit, finalizeNestedRecordInit

Part of the checker.ts modularization effort.
